[REL] Tagging release

+++ tiki-edit_templates.php
@@ -1,13 +1,19 @@
 <?php
-// (c) Copyright 2002-2011 by authors of the Tiki Wiki CMS Groupware Project
-// 
+
+// $Id: /cvsroot/tikiwiki/tiki/tiki-edit_templates.php,v 1.23 2007-10-12 07:55:27 nyloth Exp $
+
+// Copyright (c) 2002-2007, Luis Argerich, Garland Foster, Eduardo Polidor, et. al.
 // All Rights Reserved. See copyright.txt for details and a complete list of authors.
 // Licensed under the GNU LESSER GENERAL PUBLIC LICENSE. See license.txt for details.
-// $Id$
 
+// Initialization
 require_once ('tiki-setup.php');
 
-$access->check_feature('feature_view_tpl');
+if ($prefs['feature_edit_templates'] != 'y' && $prefs['feature_view_tpl'] != 'y') {
+	$smarty->assign('msg', tra("Feature disabled"));
+	$smarty->display("error.tpl");
+	die;
+}
 
 // you have to have the perm view and edit to continue:
       // if view perm is set: continue
@@ -48,7 +54,6 @@ if (isset($_REQUEST["template"])) {
 // do editing stuff only if you have the permission to:
 if ($tiki_p_edit_templates == 'y') {
 	if ((isset($_REQUEST["save"]) || isset($_REQUEST['saveTheme'])) && !empty($_REQUEST['template'])) {
-		$access->check_feature('feature_edit_templates');
 		check_ticket('edit-templates');
 		if (isset($_REQUEST['saveTheme'])) {
 			if (!empty($tikidomain)) {
@@ -76,10 +81,15 @@ if ($tiki_p_edit_templates == 'y') {
 	}
 	
 	if (isset($_REQUEST['delete']) && !empty($_REQUEST['template'])) {
-		$access->check_authenticity();
-		$file = $smarty->get_filename($_REQUEST['template']);
-		unlink($file);
-		unset($_REQUEST['template']);
+		$area = 'deltpl';
+		if ($prefs['feature_ticketlib2'] != 'y' or (isset($_POST['daconfirm']) and isset($_SESSION["ticket_$area"]))) {
+	    		key_check($area);
+			$file = $smarty->get_filename($_REQUEST['template']);
+			unlink($file);
+			unset($_REQUEST['template']);
+		} else {
+			 key_get($area);
+		}
 	}
 }
 
@@ -107,13 +117,18 @@ if (isset($_REQUEST["template"])) {
 if ($mode == 'listing') {
 	// Get templates from the templates directory
 	$local = 'styles/'.str_replace('.css', '', $prefs['style']).'/';
-	$where = array('', 'mail/', 'map/', 'modules/', $local);
+	$where = array('', 'modules/', 'mail/', 'map/', $local);
 	$files = array();
-	chdir($smarty->template_dir);
 	foreach ($where as $w) {
-		$files = array_merge($files, glob($w . '*.tpl'));
+		$h = opendir($smarty->template_dir.$w);
+		while (($file = readdir($h)) !== false) {
+			if (substr($file,-4,4) == '.tpl' && ($w != $local || !in_array($file, $files))) {
+				$files[] = $w.$file;
+			}
+		}
+		closedir ($h);
 	}
-	chdir($tikipath);
+	sort ($files);
 	$smarty->assign('files', $files);
 }
 $smarty->assign('mode', $mode);
@@ -128,3 +143,5 @@ $smarty->assign('metatag_robots', 'NOINDEX, NOFOLLOW');
 // Get templates from the templates/modules directory
 $smarty->assign('mid', 'tiki-edit_templates.tpl');
 $smarty->display("tiki.tpl");
+
+?>
